LightCurveViewer: clean up addRemoveAnnotationButton code
