+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/LightCurveViewer/d3/addRemoveAnnotationButton.spec.js
@@ -1,0 +1,7 @@
+import addRemoveAnnotationButton from './addRemoveAnnotationButton'
+
+describe('LightCurveViewer > d3 > addRemoveAnnotationButton', function () {
+  it('should exist', function () {
+    expect(addRemoveAnnotationButton).to.be.a('function')
+  })
+})
